# services/api/internal/apidocs/openapi.yaml
openapi: 3.0.3
info:
  title: 初芽记 chuyaji API
  version: "0.2.0"
  description: |
    宝宝 / 宝妈双工作台 HTTP API（Gin + SQLite）。

    认证方式：除登录、健康检查、公开附件外，请求头需携带 `Authorization: Bearer <JWT>`。
    开发模式：当服务端 `CHUYAJI_DEV_MODE=true` 时，登录 body 可使用 `{"code":"dev"}`。
servers:
  - url: http://127.0.0.1:8282
    description: 本地默认
  - url: /
    description: 相对当前主机与端口
...
paths:
...
  /api/v1/records/{id}:
    get:
      tags: [BabyRecord]
      summary: 宝宝记录详情
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/idPath"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BabyRecord"
    patch:
      tags: [BabyRecord]
      summary: 更新宝宝记录
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/idPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchBabyRecordRequest"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BabyRecord"
    delete:
      tags: [BabyRecord]
      summary: 删除宝宝记录
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/idPath"
      responses:
        "204":
          description: 已删除
...
  /api/v1/mothers/{id}:
    get:
      tags: [Mother]
      summary: 宝妈档案详情
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/idPath"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mother"
    patch:
      tags: [Mother]
      summary: 更新宝妈档案
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/idPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchMotherRequest"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mother"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    idPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
...
    BabyRecord:
      type: object
      properties:
        id:
          type: integer
          format: int64
        baby_id:
          type: integer
          format: int64
        phase:
          type: string
          enum: [prenatal, postnatal]
        record_type:
          type: string
        occurred_at:
          type: string
          format: date-time
        gestational_weeks:
          type: integer
        gestational_days:
          type: integer
        summary:
          type: string
        payload:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateBabyRecordRequest:
      type: object
      required: [phase, record_type, occurred_at]
      properties:
        phase:
          type: string
          enum: [prenatal, postnatal]
        record_type:
          type: string
        occurred_at:
          type: string
          format: date-time
        gestational_weeks:
          type: integer
        gestational_days:
          type: integer
        summary:
          type: string
        payload:
          type: object
          additionalProperties: true
    PatchBabyRecordRequest:
      allOf:
        - $ref: "#/components/schemas/CreateBabyRecordRequest"
...
    Mother:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        name:
          type: string
        birthday:
          type: string
          format: date-time
        height_cm:
          type: number
        pre_pregnancy_weight_kg:
          type: number
        blood_type:
          type: string
        allergy_history:
          type: string
        medical_history:
          type: string
        status:
          type: string
          enum: [pregnant, postpartum, parenting]
        delivery_date:
          type: string
          format: date-time
        note:
          type: string
    CreateMotherRequest:
      type: object
      properties:
        name:
          type: string
        birthday:
          type: string
          format: date-time
        height_cm:
          type: number
        pre_pregnancy_weight_kg:
          type: number
        blood_type:
          type: string
        allergy_history:
          type: string
        medical_history:
          type: string
        status:
          type: string
          enum: [pregnant, postpartum, parenting]
        delivery_date:
          type: string
          format: date-time
        note:
          type: string
    PatchMotherRequest:
      allOf:
        - $ref: "#/components/schemas/CreateMotherRequest"
...